requirements.txt
django<2.2
django-crispy-forms<1.8
django-disqus<0.6
django-background-tasks<1.3
django-ipware<2.1
requests
